23.04.25 db 24시간 최신화,price와 24hours sync
