+++ firstSolar/back/src/routes/sync.ts
@@ -12,6 +12,7 @@ const web3 = new Web3(
 
 import { abi as DexlAbi } from "../../contracts/artifacts/Dex.json";
 import price from "../priceList";
+import swapPriceSync from "../swapPriceSync";
 import { type } from "os";
 import { send } from "process";
 
@@ -108,128 +109,47 @@ router.post("/mypage", async (req, res) => {
   }
 });
 
-router.post("/swap", async (req, res) => {
-  const BtnUSDT = await db.Price.findOne({
-    where: {
-      tokenSymbol: req.body.data,
-    },
-  });
-  res.send(BtnUSDT);
-});
-router.post("/filterSwap", async (req, res) => {
-  console.log(req.body);
-  const BtnUSDT = await db.Price.findOne({
-    where: {
-      tokenSymbol: req.body.data,
-    },
-  });
-  res.send(BtnUSDT.tokenPrice);
-});
-router.get("/dateSync", async (req, res) => {
+// router.post("/swap", async (req, res) => {
+//   const BtnUSDT = await db.Price.findOne({
+//     where: {
+//       tokenSymbol: req.body.data,
+//     },
+//   });
+//   res.send(BtnUSDT);
+// });
+// router.post("/filterSwap", async (req, res) => {
+//   console.log(req.body);
+//   const BtnUSDT = await db.Price.findOne({
+//     where: {
+//       tokenSymbol: req.body.data,
+//     },
+//   });
+//   res.send(BtnUSDT.tokenPrice);
+// });
+router.get("/datesync", async (req, res) => {
   try {
-    const timeSync = async () => {
-      console.log("실행됨 1");
-      const updateTime = await db.Price.findAll();
-      const nowTime = new Date();
+    const sync = await swapPriceSync();
+    res.send(sync);
 
-      updateTime.forEach(async (pool) => {
-        console.log("실행됨 2");
-
-        const lastUpdateTime = new Date(pool.dataValues.updatedAt);
-        // DB에서 불러온 각 Pool 인스턴스의 업데이트 시간을 가져와 'lastUpdateTime' 변수에 할당
-
-        const timeDiffInMs = nowTime.getTime() - lastUpdateTime.getTime();
-        // 현재 시간과 'lastUpdateTime'의 차이(밀리초)를 계산하여 'timeDiffInMs' 변수에 할당
-
-        const timeDiffInHours = timeDiffInMs / (1000 * 60 * 60);
-        // 'timeDiffInMs'를 시간 단위로 변환하여 'timeDiffInHours' 변수에 할당
-
-        // 'timeDiffInHours'가 24시간을 초과 하는지 여부에 따라 메세지 출력
-
-        const DbList = await db.Price.findAll();
-
-        if (DbList.length <= 0) {
-          console.log("실행됨 3");
-
-          // 24시간이 지났고, db에 값이 없을때
-          let priceDataList = await price();
-          for (let i = 0; i < priceDataList.length; i++) {
-            console.log("실행됨 4");
-
-            await db.Price.create({
-              tokenSymbol: priceDataList[i].symbolData,
-              tokenId: priceDataList[i].IdData,
-              tokenSlug: priceDataList[i].slugData,
-              ConvertToUSDT: priceDataList[i].convertUSDT,
-              ConvertToETH: priceDataList[i].convertETH,
-              ConvertToBNB: priceDataList[i].convertBNB,
-              tokenPrice: priceDataList[i].tokenPrice,
-            });
-          }
-        } else {
-          console.log("실행됨 5");
-
-          if (timeDiffInHours > 24) {
-            console.log("24시간 이상 지났습니다.");
-            let priceDataList = await price();
-            let priceDB = await db.Price.findAll();
-            for (let i = 0; i < priceDataList.length; i++) {
-              console.log("실행됨 6");
-
-              await db.Price.update(
-                {
-                  ConvertToUSDT: priceDB[i].convertUSDT,
-                  ConvertToETH: priceDB[i].convertETH,
-                  ConvertToBNB: priceDB[i].convertBNB,
-                  tokenPrice: priceDB[i].tokenPrice,
-                },
-                {
-                  where: {
-                    tokenSymbol: priceDataList[i].symbolData,
-                    tokenId: priceDataList[i].IdData,
-                    tokenSlug: priceDataList[i].slugData,
-                  },
-                }
-              );
-            }
-            priceDB = await db.Price.findAll();
-            console.log(priceDB);
-            return;
-          } else {
-            let priceDB = await db.Price.findAll();
-            // console.log("priceDB", priceDB);
-            // console.log("24시간 이내입니다.");
-            return;
-          }
-
-          // 24시간이 지났고, db에 지난 값들이 있을때
-        }
-      });
-    };
-    timeSync();
     const now = new Date();
-
     const nextUpdate = new Date(
       now.getFullYear(),
       now.getMonth(),
       now.getDate() + 1,
       9
     );
-
     // 다음 업데이트 시간까지 남은 시간(밀리초) 계산
     const timeUntilNextUpdate = nextUpdate.getTime() - now.getTime();
-
     // setInterval() 메서드로 주기적으로 checkUpdateTime() 함수 실행
-    setInterval(() => {
-      timeSync();
-    }, 60 * 1000);
-
+    setInterval(async () => {
+      const repeatedSync = await swapPriceSync();
+      console.log("repeatedSync", repeatedSync);
+    }, 24 * 60 * 60 * 1000);
     // 첫 번째 업데이트는 다음 업데이트 시간까지 남은 시간 후에 실행
-    setTimeout(() => {
-      timeSync();
+    setTimeout(async () => {
+      const firstUpdateSync = await swapPriceSync();
+      console.log("firstUpdateSync", firstUpdateSync);
     }, timeUntilNextUpdate);
-
-    res.end();
   } catch (error) {
     console.log(error);
   }
